Fix linting issues
This provides a huge swath of linting fixes.

None of these are intended to change any functionality, although without
tests we just have to hope.

There is one linting issue which could not be addressed without
meaningful re-architecture, and that is property re-assignment.  For now
that rule is just being disabled so we can address that separately as
part of a larger rework.

Issue #3