Fork ts-demuxer
TS-Demuxer is an excellent project by Logan Kearsley (@gliese1337) which
ports an MPEG-TS demuxer from C++ into TypeScript.

You can view that project on npm at:
https://www.npmjs.com/package/ts-demuxer

Or on GitHub at:
https://github.com/gliese1337/HLS.js/tree/master/demuxer

The project is lightly maintained and Logan does engage with issues,
however because demuxing is such a critical component of TV Kitchen we
decided it would be good to maintain a version directly.

There are some existing known bugs in the current implementation which
need to be resolved before TV Kitchen can be used in production, such
as:

- https://github.com/gliese1337/HLS.js/issues/15
- https://github.com/gliese1337/HLS.js/issues/16

As part of this fork we are going to end up refactoring the project
a bit to bring it up to more strict linting standards, and may also
explore some architectural changes as well.

Changes made in this repository are also being released under a new
license, which can be viewed at:

https://github.com/tvkitchen/utilities/blob/main/LICENSE.md

Issue #3