Remove len and offset parameters
The `process` method used to take in an offset parameter as well as
a length parameter.  This made the management of data much more
complicated without any clear benefit.

By removing these parameters and instead simply taking in a buffer of
data to process we dramatically simplify the API as well as the
implementation of `process`.

Issue #8