Simplify process
The previous process method had a lot of moving parts and this is an
attempt to dramatically simplify it while also avoiding the loss of
data.

We no longer care about reporting back corrupt packets, but instead the
demuxer will attempt to continue until there is no data left.

We no longer have a concept of `leftover` but instead just have
a concept of `unprocessed`. This is a buffer that is constantly updated
as data is passed to the demuxer.

The demuxer no longer exists if there is an error in the data, but
instead skips bytes deemed useless by the demuxPacket helper methods.

Issue #8