Allow pts rollover
The MPEG-TS pts field can roll over after a stream has been on for
a sufficiently long time (e.g. a week or so).

There was a failsafe built into the demuxer that would prevent PTS from
ever decreasing -- this was because normally PTS is not supposed to
decrease (outside of rollovers).

This change now assumes that the data stream will be passed in order in
order to maintain that PTS condition.

Issue #18